// packages/foundation-web/src/components/table/preset/data-table/component/table-row-action/type.ts
import { EnumFormMode, UUID } from '@kilnonedre/foundation'
import { Row, Table } from '@tanstack/react-table'

export interface ConfigProp<T> {
  row: Row<T>
  table: Table<T>
  toEdit?: (_id: string) => void
  toDelete?: (_id: string) => void
  toAudit?: (_id: string) => void
}

export interface FormDialogState {
  mode: EnumFormMode
  id?: UUID
}